+++ src/main/java/rbasamoyai/createbigcannons/datagen/CBCDatagenCommon.java
@@ -1,10 +1,14 @@
 package rbasamoyai.createbigcannons.datagen;
 
+import rbasamoyai.createbigcannons.CBCTags;
+
 public class CBCDatagenCommon {
 
 	public static final CBCDatagenPlatform PLATFORM = CBCDatagenPlatform.getPlatform(System.getProperty("createbigcannons.datagen.platform"));
 	public static final int FLUID_MULTIPLIER = PLATFORM.fluidMultiplier();
 
-	public static void init() {}
+	public static void init() {
+        CBCTags.register();
+    }
 
 }