+++ CHANGELOG.md
@@ -50,6 +50,7 @@ Fixed:
 - Fixed sliding breech and quick-firing breech breechblock rendering
 - Fixed cannon cast error fluid name display
 - Fixed creative autocannon ammo container screen display title when accessed in block form
+- Fixed dispenser behavior of molten metals
 
 ## [5.10.2] - 2026-01-17